% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{circos_plot}
\alias{circos_plot}
\title{Draw final circle plot}
\usage{
circos_plot(
  expression,
  cnv,
  variant,
  arriba,
  bed,
  SNP_max_rad = 1.145,
  SNP_min_rad = 1,
  exp_max_rad = 0.97,
  exp_min_rad = 0.8,
  CNV_max_rad = 0.75,
  CNV_min_rad = 0.4,
  fusion_max_rad = 0.35,
  CNV_cutoff = 1.5
)
}
\description{
Draw final circle plot
}
