% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fast_map}
\alias{fast_map}
\title{Input = df with samples as rownames and features as cols
Annotation = vector of annotation (in the same order as input's rownames!!)
anno_pal = viridis_pal, hue_pal, grey_pal...}
\usage{
fast_map(
  df,
  top_features_p = 1,
  center_scale = T,
  heat_colors = colorRampPalette(c("#3361A5", "#248AF3", "#14B3FF", "#88CEEF",
    "#C1D5DC", "#EAD397", "#FDB31A", "#E42A2A", "#A31D1D"))(100),
  anno_pal = viridis_pal(),
  Zlim = 3,
  method = "ward.D2",
  metric = "euclidean",
  title = "title",
  Data_input = "Z-score expression",
  anno_1 = NULL,
  anno_1_name = NULL,
  anno_1_cols = NULL,
  anno_2 = NULL,
  anno_2_name = NULL,
  anno_2_cols = NULL,
  cluster_rows = T,
  cluster_cols = T,
  reorder_rows = T,
  reorder_cols = T,
  row_names = T,
  col_names = T,
  row_fontsize = 12,
  col_fontsize = 12
)
}
\description{
Input = df with samples as rownames and features as cols
Annotation = vector of annotation (in the same order as input's rownames!!)
anno_pal = viridis_pal, hue_pal, grey_pal...
}
