% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{tx2gene}
\alias{tx2gene}
\title{Global parameters}
\format{
An object of class \code{data.frame} with 237012 rows and 2 columns.
}
\usage{
tx2gene
}
\description{
Global parameters
}
\keyword{datasets}
